Factor out path_utils.rs
